Add __fp16 for half-precision floating-point types
In order to provide the broadest level of compatibility across language versions
and compilers in libraries and programs, we adopt the `__fp16` rather than `_Float16`
as half-precision floating-point types. For example in the `printf` case, there is
no format specifier for the C language extension type half-precision floating-point.
In the case of passing a `_Float16` to `%f` or `%d` the value will be ignored and
silently fail, and no value will be formatted. In the case of `__fp16`, there is an
implicit cast to double, therefore the value will be formatted. So if you pass
`_Float16` to printf, you won't be able to print it, as there is no printf format code
for half-precision floating-point types. If you pass __fp16 to printf, it will print
fine, as it will promote to double.